move qwen36 moe cli modules into directory
