+++ crates/runner/src/qwen36_moe/geom.rs
