+++ crates/runner/src/qwen36_moe/host.rs
@@ -2,7 +2,7 @@ use anyhow::{anyhow, Context, Result};
 use model_store::BakedStore;
 use qwen36_moe::config::TextConfig;
 
-use crate::qwen36_moe_layers::QWEN36_MOE_INT4_GROUP_SIZE;
+use crate::qwen36_moe_cli::layers::QWEN36_MOE_INT4_GROUP_SIZE;
 use crate::qwen36_moe_logits::dequant_int4_to_bf16_bytes;
 use crate::qwen36_moe_types::MultiLayerGeom;
