+++ crates/runner/src/qwen36_moe/prefetch.rs
