+++ crates/runner/src/qwen36_moe/session.rs
@@ -3,8 +3,8 @@ use gpu_hal::{GpuBuffer, ScalarType};
 use model_store::BakedStore;
 use qwen36_moe::config::TextConfig;
 
-use crate::qwen36_moe_host::{host_load_bytes, load_lm_head_bf16};
-use crate::qwen36_moe_layers::load_to_gpu;
+use crate::qwen36_moe_cli::host::{host_load_bytes, load_lm_head_bf16};
+use crate::qwen36_moe_cli::layers::load_to_gpu;
 use crate::qwen36_moe_types::{LayerBuffers, MtpLayerBuffers, MultiLayerGeom};
 
 const MIB: f64 = (1024 * 1024) as f64;
@@ -38,7 +38,7 @@ pub(crate) fn prepare_decode_session(
     layers: &mut Vec<LayerBuffers>,
 ) -> Result<Qwen36DecodeSession> {
     let mtp_buffers_opt = if speculative_decode {
-        match crate::qwen36_moe_mtp_loader::load_mtp_buffers(store, ordinal, geom, kv_max_t)
+        match crate::qwen36_moe_cli::mtp_loader::load_mtp_buffers(store, ordinal, geom, kv_max_t)
             .context("load MTP head from bake")?
         {
             Some(mtp) => {
