+++ crates/runner/src/qwen36_moe/timing.rs
